Move server creation function to api_server module
This was introduced in the 1.9.51 SDK.
