+++ AppServer/google/appengine/tools/devappserver2/api_server.py
@@ -20,6 +20,9 @@ The Remote API protocol is used for communication.
 """
 
 
+import errno
+import getpass
+import itertools
 import logging
 import os
 import pickle
@@ -65,6 +68,7 @@ from google.appengine.ext.cloudstorage import stub_dispatcher as gcs_dispatcher
 from google.appengine.ext.remote_api import remote_api_pb
 from google.appengine.ext.remote_api import remote_api_services
 from google.appengine.runtime import apiproxy_errors
+from google.appengine.tools.devappserver2 import login
 from google.appengine.tools.devappserver2 import wsgi_server
 
 # AppScale
@@ -207,6 +211,186 @@ class APIServer(wsgi_server.WsgiServer):
     else:
       start_response('405 Method Not Allowed')
       return []
+
+
+def create_api_server(request_info, storage_path, options, configuration):
+  """Creates an API server.
+
+  Args:
+    request_info: An apiproxy_stub.RequestInfo instance used by the stubs to
+      lookup information about the request associated with an API call.
+    storage_path: A string directory for storing API stub data.
+    options: An instance of argparse.Namespace containing command line flags.
+    configuration: An instance of
+      application_configuration.ApplicationConfiguration for configuring API
+      stub settings.
+
+  Returns:
+    An instance of APIServer.
+  """
+  datastore_path = options.datastore_path or os.path.join(
+      storage_path, 'datastore.db')
+  logs_path = options.logs_path or os.path.join(storage_path, 'logs.db')
+  search_index_path = options.search_indexes_path or os.path.join(
+      storage_path, 'search_indexes')
+  prospective_search_path = options.prospective_search_path or os.path.join(
+      storage_path, 'prospective-search')
+  blobstore_path = options.blobstore_path or os.path.join(
+      storage_path, 'blobs')
+
+  if options.clear_datastore:
+    _clear_datastore_storage(datastore_path)
+
+  if options.clear_prospective_search:
+    _clear_prospective_search_storage(prospective_search_path)
+
+  if options.clear_search_indexes:
+    _clear_search_indexes_storage(search_index_path)
+  if options.auto_id_policy == datastore_stub_util.SEQUENTIAL:
+    logging.warn("--auto_id_policy='sequential' is deprecated. This option "
+                 "will be removed in a future release.")
+
+  application_address = '%s' % options.host
+  if options.port and options.port != 80:
+    application_address += ':' + str(options.port)
+
+  user_login_url = '/%s?%s=%%s' % (
+      login.LOGIN_URL_RELATIVE, login.CONTINUE_PARAM)
+  user_logout_url = '%s&%s=%s' % (
+      user_login_url, login.ACTION_PARAM, login.LOGOUT_ACTION)
+
+  if options.datastore_consistency_policy == 'time':
+    consistency = datastore_stub_util.TimeBasedHRConsistencyPolicy()
+  elif options.datastore_consistency_policy == 'random':
+    consistency = datastore_stub_util.PseudoRandomHRConsistencyPolicy()
+  elif options.datastore_consistency_policy == 'consistent':
+    consistency = datastore_stub_util.PseudoRandomHRConsistencyPolicy(1.0)
+  else:
+    assert 0, ('unknown consistency policy: %r' %
+               options.datastore_consistency_policy)
+
+  maybe_convert_datastore_file_stub_data_to_sqlite(
+      configuration.app_id, datastore_path)
+  setup_stubs(
+      request_data=request_info,
+      app_id=configuration.app_id,
+      application_root=configuration.modules[0].application_root,
+      # The "trusted" flag is only relevant for Google administrative
+      # applications.
+      trusted=getattr(options, 'trusted', False),
+      blobstore_path=blobstore_path,
+      datastore_path=datastore_path,
+      datastore_consistency=consistency,
+      datastore_require_indexes=options.require_indexes,
+      datastore_auto_id_policy=options.auto_id_policy,
+      images_host_prefix='http://%s' % application_address,
+      logs_path=logs_path,
+      mail_smtp_host=options.smtp_host,
+      mail_smtp_port=options.smtp_port,
+      mail_smtp_user=options.smtp_user,
+      mail_smtp_password=options.smtp_password,
+      mail_enable_sendmail=options.enable_sendmail,
+      mail_show_mail_body=options.show_mail_body,
+      matcher_prospective_search_path=prospective_search_path,
+      search_index_path=search_index_path,
+      taskqueue_auto_run_tasks=options.enable_task_running,
+      taskqueue_default_http_server=application_address,
+      user_login_url=user_login_url,
+      user_logout_url=user_logout_url,
+      default_gcs_bucket_name=options.default_gcs_bucket_name,
+      uaserver_path=options.uaserver_path,
+      xmpp_path=options.xmpp_path)
+
+  # The APIServer must bind to localhost because that is what the runtime
+  # instances talk to.
+  return APIServer('localhost', options.api_port, configuration.app_id)
+
+
+def _clear_datastore_storage(datastore_path):
+  """Delete the datastore storage file at the given path."""
+  # lexists() returns True for broken symlinks, where exists() returns False.
+  if os.path.lexists(datastore_path):
+    try:
+      os.remove(datastore_path)
+    except OSError, err:
+      logging.warning(
+          'Failed to remove datastore file %r: %s', datastore_path, err)
+
+
+def _clear_prospective_search_storage(prospective_search_path):
+  """Delete the perspective search storage file at the given path."""
+  # lexists() returns True for broken symlinks, where exists() returns False.
+  if os.path.lexists(prospective_search_path):
+    try:
+      os.remove(prospective_search_path)
+    except OSError, err:
+      logging.warning(
+          'Failed to remove prospective search file %r: %s',
+          prospective_search_path, err)
+
+
+def _clear_search_indexes_storage(search_index_path):
+  """Delete the search indexes storage file at the given path."""
+  # lexists() returns True for broken symlinks, where exists() returns False.
+  if os.path.lexists(search_index_path):
+    try:
+      os.remove(search_index_path)
+    except OSError, err:
+      logging.warning(
+          'Failed to remove search indexes file %r: %s', search_index_path, err)
+
+
+def get_storage_path(path, app_id):
+  """Returns a path to the directory where stub data can be stored."""
+  _, _, app_id = app_id.replace(':', '_').rpartition('~')
+  if path is None:
+    for path in _generate_storage_paths(app_id):
+      try:
+        os.mkdir(path, 0700)
+      except OSError, err:
+        if err.errno == errno.EEXIST:
+          # Check that the directory is only accessable by the current user to
+          # protect against an attacker creating the directory in advance in
+          # order to access any created files. Windows has per-user temporary
+          # directories and st_mode does not include per-user permission
+          # information so assume that it is safe.
+          if sys.platform == 'win32' or (
+              (os.stat(path).st_mode & 0777) == 0700 and os.path.isdir(path)):
+            return path
+          else:
+            continue
+        raise
+      else:
+        return path
+  elif not os.path.exists(path):
+    os.mkdir(path)
+    return path
+  elif not os.path.isdir(path):
+    raise IOError('the given storage path %r is a file, a directory was '
+                  'expected' % path)
+  else:
+    return path
+
+
+def _generate_storage_paths(app_id):
+  """Yield an infinite sequence of possible storage paths."""
+  if sys.platform == 'win32':
+    # The temp directory is per-user on Windows so there is no reason to add
+    # the username to the generated directory name.
+    user_format = ''
+  else:
+    try:
+      user_name = getpass.getuser()
+    except Exception:  # pylint: disable=broad-except
+      # The possible set of exceptions is not documented.
+      user_format = ''
+    else:
+      user_format = '.%s' % user_name
+
+  tempdir = tempfile.gettempdir()
+  yield os.path.join(tempdir, 'appengine.%s%s' % (app_id, user_format))
+  for i in itertools.count(1):
+    yield os.path.join(tempdir, 'appengine.%s%s.%d' % (app_id, user_format, i))
 
 
 def setup_stubs(
