+++ AppServer/google/appengine/tools/devappserver2/api_server_test.py
@@ -18,11 +18,18 @@
 
 
 import cStringIO
+import getpass
+import itertools
+import os
 import pickle
+import sys
 import tempfile
 import unittest
 import urllib
 import wsgiref.util
+
+import google
+import mox
 
 from google.appengine.api import apiproxy_stub
 from google.appengine.api import apiproxy_stub_map
@@ -188,6 +195,116 @@ class TestAPIServer(wsgi_test_utils.WSGITestCase):
 
     self._assert_remote_call(
         expected_remote_response, urlfetch_request, 'urlfetch', 'Fetch')
+
+
+class GetStoragePathTest(unittest.TestCase):
+  """Tests for api_server.get_storage_path."""
+
+  def setUp(self):
+    self.mox = mox.Mox()
+    self.mox.StubOutWithMock(api_server, '_generate_storage_paths')
+
+  def tearDown(self):
+    self.mox.UnsetStubs()
+
+  def test_no_path_given_directory_does_not_exist(self):
+    path = tempfile.mkdtemp()
+    os.rmdir(path)
+    api_server._generate_storage_paths('example.com_myapp').AndReturn([path])
+
+    self.mox.ReplayAll()
+    self.assertEqual(
+        path, api_server.get_storage_path(None, 'dev~example.com:myapp'))
+    self.mox.VerifyAll()
+    self.assertTrue(os.path.isdir(path))
+
+  def test_no_path_given_directory_exists(self):
+    path1 = tempfile.mkdtemp()
+    os.chmod(path1, 0777)
+    path2 = tempfile.mkdtemp()  # Made with mode 0700.
+
+    api_server._generate_storage_paths('example.com_myapp').AndReturn(
+        [path1, path2])
+
+    self.mox.ReplayAll()
+    if sys.platform == 'win32':
+      expected_path = path1
+    else:
+      expected_path = path2
+    self.assertEqual(
+        expected_path,
+        api_server.get_storage_path(None, 'dev~example.com:myapp'))
+    self.mox.VerifyAll()
+
+  def test_path_given_does_not_exist(self):
+    path = tempfile.mkdtemp()
+    os.rmdir(path)
+
+    self.assertEqual(
+        path, api_server.get_storage_path(path, 'dev~example.com:myapp'))
+    self.assertTrue(os.path.isdir(path))
+
+  def test_path_given_not_directory(self):
+    _, path = tempfile.mkstemp()
+
+    self.assertRaises(
+        IOError, api_server.get_storage_path, path, 'dev~example.com:myapp')
+
+  def test_path_given_exists(self):
+    path = tempfile.mkdtemp()
+
+    self.assertEqual(
+        path, api_server.get_storage_path(path, 'dev~example.com:myapp'))
+
+
+class GenerateStoragePathsTest(unittest.TestCase):
+  """Tests for api_server._generate_storage_paths."""
+
+  def setUp(self):
+    self.mox = mox.Mox()
+    self.mox.StubOutWithMock(getpass, 'getuser')
+    self.mox.StubOutWithMock(tempfile, 'gettempdir')
+
+  def tearDown(self):
+    self.mox.UnsetStubs()
+
+  @unittest.skipUnless(sys.platform.startswith('win'), 'Windows only')
+  def test_windows(self):
+    tempfile.gettempdir().AndReturn('/tmp')
+
+    self.mox.ReplayAll()
+    self.assertEqual(
+        [os.path.join('/tmp', 'appengine.myapp'),
+         os.path.join('/tmp', 'appengine.myapp.1'),
+         os.path.join('/tmp', 'appengine.myapp.2')],
+        list(itertools.islice(api_server._generate_storage_paths('myapp'), 3)))
+    self.mox.VerifyAll()
+
+  @unittest.skipIf(sys.platform.startswith('win'), 'not on Windows')
+  def test_working_getuser(self):
+    getpass.getuser().AndReturn('johndoe')
+    tempfile.gettempdir().AndReturn('/tmp')
+
+    self.mox.ReplayAll()
+    self.assertEqual(
+        [os.path.join('/tmp', 'appengine.myapp.johndoe'),
+         os.path.join('/tmp', 'appengine.myapp.johndoe.1'),
+         os.path.join('/tmp', 'appengine.myapp.johndoe.2')],
+        list(itertools.islice(api_server._generate_storage_paths('myapp'), 3)))
+    self.mox.VerifyAll()
+
+  @unittest.skipIf(sys.platform.startswith('win'), 'not on Windows')
+  def test_broken_getuser(self):
+    getpass.getuser().AndRaise(Exception())
+    tempfile.gettempdir().AndReturn('/tmp')
+
+    self.mox.ReplayAll()
+    self.assertEqual(
+        [os.path.join('/tmp', 'appengine.myapp'),
+         os.path.join('/tmp', 'appengine.myapp.1'),
+         os.path.join('/tmp', 'appengine.myapp.2')],
+        list(itertools.islice(api_server._generate_storage_paths('myapp'), 3)))
+    self.mox.VerifyAll()
 
 
 if __name__ == '__main__':
